##############第一版提交####################### 新增: 修改: 修复: 缺陷:
待完成:#去掉了发布商品时照片和文本非法检(UserController677-683)
      点击菜单时会偶然出现白屏
      ******后台界面******

+++ src/main/java/com/sss/dao/OrderFormMapper.java
@@ -5,13 +5,13 @@ import com.sss.pojo.OrderForm;
 import java.util.List;
 
 public interface OrderFormMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(OrderForm record);
 
     int insertSelective(OrderForm record);
 
-    OrderForm selectByPrimaryKey(Integer id);
+    OrderForm selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(OrderForm record);
